requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.7.4
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cymem==2.0.5
discord-py-slash-command==1.1.0
discord.py==1.7.1
en-core-web-sm==2.3.1
idna==2.10
more-itertools==8.7.0
multidict==5.1.0
murmurhash==1.0.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
numpy==1.20.2
ordered-set==3.1.1
ordered-set-stubs==0.1.3
pathspec==0.8.1
Pillow==8.2.0
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.5
profanity-filter==1.3.3
protobuf==3.15.8
pydantic==1.8.1
python-colored-print==1.0.0rc1
python-dotenv==0.17.0
redis==3.5.3
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.15.100
six==1.15.0
soupsieve==2.2.1
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
toml==0.10.2
tomlkit==0.5.11
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
yarl==1.6.3
